Moving node_types and editor functions to their own modules
