+++ ix/chains/management/commands/import_langchain.py
@@ -1,6 +1,6 @@
 from django.core.management.base import BaseCommand
 
-from ix.api.chains.types import NodeTypeField
+from ix.api.components.types import NodeTypeField
 from ix.api.chains.types import NodeType as NodeTypePydantic
 from ix.chains.fixture_src.agent_interaction import AGENT_INTERACTION_CHAINS
