+++ ix/chains/tests/test_config.py
@@ -4,7 +4,7 @@ from typing import Literal, Optional
 
 import pytest
 from pydantic import BaseModel
-from ix.api.chains.types import NodeTypeField, InputType
+from ix.api.components.types import NodeTypeField, InputType
 
 
 class ChoicesEnum(str, Enum):
